lib: Merge qunit-config.js into qunit-tests.js
This avoids the awkward `window.Qunit` handover and just uses simple
static imports. It also avoids the dynamic require().

Both files are small enough, and nothing else requires qunit-config.js
anyway.
