+++ pkg/lib/qunit-tests.js
@@ -23,12 +23,6 @@ import QUnit from "qunit/qunit/qunit.js";
 import qunitTap from "qunit-tap/lib/qunit-tap.js";
 import "qunit/qunit/qunit.css";
 
-/* QUnit needs to have 'window' as 'this' in order to load */
-window.QUnit = QUnit;
-window.qunitTap = qunitTap;
-
-require("./qunit-config.js");
-
 QUnit.mock_info = async key => {
     const response = await fetch(`http://${window.location.hostname}:${window.location.port}/mock/info`);
     return (await response.json())[key];
@@ -41,5 +35,54 @@ if (await QUnit.mock_info("pybridge"))
     QUnit.test.skipWithPybridge = QUnit.test.skip;
 else
     QUnit.test.skipWithPybridge = QUnit.test;
+
+/* Always use explicit start */
+QUnit.config.autostart = false;
+
+let qunit_started = false;
+
+QUnit.moduleStart(function() {
+    qunit_started = true;
+});
+
+QUnit.done(function() {
+    /*
+     * QUnit-Tap writes the summary line right after this function returns.
+     * Delay printing the end marker until after that summary is out.
+     */
+    window.setTimeout(function () {
+        console.log("cockpittest-tap-done");
+    }, 0);
+});
+/*
+ * Now initialize qunit-tap
+ *
+ * When not running under a tap driver this stuff will just show up in
+ * the console. We print out a special canary at the end of the tests
+ * so that the tap driver can know when the testing is done.
+ *
+ * In addition double check for a test file that doesn't properly call
+ * QUnit.start() after its done setting up its tests.
+ *
+ * We also want to insert the current test name into all tap lines.
+ */
+const tap_regex = /^((not )?ok [0-9]+ (- )?)(.*)$/;
+qunitTap(QUnit, function() {
+    if (arguments.length == 1 && QUnit.config.current) {
+        const match = tap_regex.exec(arguments[0]);
+        if (match) {
+            console.log(match[1] + QUnit.config.current.testName + ": " + match[4]);
+            return;
+        }
+    }
+    console.log.apply(console, arguments);
+});
+
+window.setTimeout(function() {
+    if (!qunit_started) {
+        console.log("QUnit not started by test");
+        console.log("cockpittest-tap-error");
+    }
+}, 20000);
 
 export default QUnit;
